Web Services Access
Previous version allowed for 6 codes to be set manually through the App.  This change introduces the concept of "Code7" a code that is not set through the App but remotely through web services, whose user name and notifications settings are set locally.  API to set the code is:  https://<webservicesURL>/setCode7/<code>